refactor :: 도메인간 기능 이전 및 파일 구조 수정
