+++ danuri-rest/src/main/kotlin/org/aing/danurirest/domain/admin/controller/AdminDeviceController.kt
@@ -1,18 +1,14 @@
-package org.aing.danurirest.domain.auth.admin.controller
+package org.aing.danurirest.domain.admin.controller
 
-import org.aing.danurirest.domain.auth.admin.dto.DeviceResponse
-import org.aing.danurirest.domain.auth.admin.dto.DeviceSignInRequest
-import org.aing.danurirest.domain.auth.admin.dto.RegisterDeviceRequest
-import org.aing.danurirest.domain.auth.admin.dto.SignInResponse
-import org.aing.danurirest.domain.auth.admin.dto.UpdateDeviceRequest
-import org.aing.danurirest.domain.auth.admin.usecase.DeleteDeviceUsecase
-import org.aing.danurirest.domain.auth.admin.usecase.DeviceSignInUsecase
-import org.aing.danurirest.domain.auth.admin.usecase.GetAdminCompanyDevicesUsecase
-import org.aing.danurirest.domain.auth.admin.usecase.GetAdminCompanyIdUsecase
-import org.aing.danurirest.domain.auth.admin.usecase.GetDeviceUsecase
-import org.aing.danurirest.domain.auth.admin.usecase.GetDevicesBySpaceUsecase
-import org.aing.danurirest.domain.auth.admin.usecase.RegisterDeviceUsecase
-import org.aing.danurirest.domain.auth.admin.usecase.UpdateDeviceUsecase
+import org.aing.danurirest.domain.admin.dto.DeviceResponse
+import org.aing.danurirest.domain.admin.dto.RegisterDeviceRequest
+import org.aing.danurirest.domain.admin.dto.UpdateDeviceRequest
+import org.aing.danurirest.domain.admin.usecase.DeleteDeviceUsecase
+import org.aing.danurirest.domain.admin.usecase.GetAdminCompanyDevicesUsecase
+import org.aing.danurirest.domain.admin.usecase.GetDeviceUsecase
+import org.aing.danurirest.domain.admin.usecase.GetDevicesBySpaceUsecase
+import org.aing.danurirest.domain.admin.usecase.RegisterDeviceUsecase
+import org.aing.danurirest.domain.admin.usecase.UpdateDeviceUsecase
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.DeleteMapping
 import org.springframework.web.bind.annotation.GetMapping
@@ -28,13 +24,11 @@ import java.util.UUID
 @RequestMapping("/admin/devices")
 class AdminDeviceController(
     private val registerDeviceUsecase: RegisterDeviceUsecase,
-    private val deviceSignInUsecase: DeviceSignInUsecase,
     private val getDeviceUsecase: GetDeviceUsecase,
     private val getAdminCompanyDevicesUsecase: GetAdminCompanyDevicesUsecase,
     private val getDevicesBySpaceUsecase: GetDevicesBySpaceUsecase,
     private val updateDeviceUsecase: UpdateDeviceUsecase,
     private val deleteDeviceUsecase: DeleteDeviceUsecase,
-    private val getAdminCompanyIdUsecase: GetAdminCompanyIdUsecase
 ) {
     @PostMapping
     fun registerDevice(
@@ -43,25 +37,17 @@ class AdminDeviceController(
         registerDeviceUsecase.execute(registerDeviceRequest).run {
             ResponseEntity.noContent().build()
         }
-
-    @PostMapping("/token")
-    fun generateDeviceToken(
-        @RequestBody deviceSignInRequest: DeviceSignInRequest,
-    ): ResponseEntity<SignInResponse> =
-        deviceSignInUsecase.execute(deviceSignInRequest).run {
-            ResponseEntity.ok(this)
-        }
         
     @GetMapping("/{deviceId}")
     fun getDevice(
         @PathVariable deviceId: UUID,
-    ): ResponseEntity<DeviceResponse> = 
+    ): ResponseEntity<DeviceResponse> =
         getDeviceUsecase.execute(deviceId).run {
             ResponseEntity.ok(this)
         }
         
     @GetMapping
-    fun getAllDevices(): ResponseEntity<List<DeviceResponse>> = 
+    fun getAllDevices(): ResponseEntity<List<DeviceResponse>> =
         getAdminCompanyDevicesUsecase.execute().run {
             ResponseEntity.ok(this)
         }
@@ -69,7 +55,7 @@ class AdminDeviceController(
     @GetMapping("/space/{spaceId}")
     fun getDevicesBySpace(
         @PathVariable spaceId: UUID,
-    ): ResponseEntity<List<DeviceResponse>> = 
+    ): ResponseEntity<List<DeviceResponse>> =
         getDevicesBySpaceUsecase.execute(spaceId).run {
             ResponseEntity.ok(this)
         }
@@ -78,7 +64,7 @@ class AdminDeviceController(
     fun updateDevice(
         @PathVariable deviceId: UUID,
         @RequestBody updateDeviceRequest: UpdateDeviceRequest,
-    ): ResponseEntity<DeviceResponse> = 
+    ): ResponseEntity<DeviceResponse> =
         updateDeviceUsecase.execute(deviceId, updateDeviceRequest).run {
             ResponseEntity.ok(this)
         }
