+++ danuri-rest/src/main/kotlin/org/aing/danurirest/domain/admin/dto/DeviceResponse.kt
@@ -1,4 +1,4 @@
-package org.aing.danurirest.domain.auth.admin.dto
+package org.aing.danurirest.domain.admin.dto
 
 import org.aing.danuridomain.persistence.device.entity.Device
 import java.time.LocalDateTime
