+++ danuri-rest/src/main/kotlin/org/aing/danurirest/domain/admin/dto/RegisterDeviceRequest.kt
@@ -1,0 +1,8 @@
+package org.aing.danurirest.domain.admin.dto
+
+import java.util.UUID
+
+data class RegisterDeviceRequest(
+    val deviceId: UUID,
+    val spaceId: UUID
+)
