+++ danuri-rest/src/main/kotlin/org/aing/danurirest/domain/admin/usecase/DeleteDeviceUsecase.kt
@@ -1,6 +1,7 @@
-package org.aing.danurirest.domain.auth.admin.usecase
+package org.aing.danurirest.domain.admin.usecase
 
 import org.aing.danuridomain.persistence.device.repository.DeviceRepository
+import org.aing.danurirest.domain.auth.admin.usecase.GetAdminCompanyIdUsecase
 import org.aing.danurirest.global.exception.CustomException
 import org.aing.danurirest.global.exception.enums.CustomErrorCode
 import org.springframework.stereotype.Service
