+++ danuri-rest/src/main/kotlin/org/aing/danurirest/domain/admin/usecase/GetDeviceUsecase.kt
@@ -1,7 +1,8 @@
-package org.aing.danurirest.domain.auth.admin.usecase
+package org.aing.danurirest.domain.admin.usecase
 
 import org.aing.danuridomain.persistence.device.repository.DeviceRepository
-import org.aing.danurirest.domain.auth.admin.dto.DeviceResponse
+import org.aing.danurirest.domain.admin.dto.DeviceResponse
+import org.aing.danurirest.domain.auth.admin.usecase.GetAdminCompanyIdUsecase
 import org.aing.danurirest.global.exception.CustomException
 import org.aing.danurirest.global.exception.enums.CustomErrorCode
 import org.springframework.stereotype.Service
