--- danuri-rest/src/main/kotlin/org/aing/danurirest/domain/auth/admin/dto/DeviceSignInRequest.kt
@@ -1,7 +1,0 @@
-package org.aing.danurirest.domain.auth.admin.dto
-
-import java.util.UUID
-
-data class DeviceSignInRequest(
-    val deviceId: UUID
-)
