+++ danuri-rest/src/main/kotlin/org/aing/danurirest/domain/auth/admin/usecase/SignInUsecase.kt
@@ -4,7 +4,7 @@ import io.github.bucket4j.Bucket
 import org.aing.danuridomain.persistence.admin.entity.Admin
 import org.aing.danuridomain.persistence.admin.repository.AdminRepository
 import org.aing.danurirest.domain.auth.admin.dto.SignInRequest
-import org.aing.danurirest.domain.auth.admin.dto.SignInResponse
+import org.aing.danurirest.domain.auth.common.dto.SignInResponse
 import org.aing.danurirest.global.exception.CustomException
 import org.aing.danurirest.global.exception.enums.CustomErrorCode
 import org.aing.danurirest.global.security.jwt.JwtProvider
