+++ danuri-rest/src/main/kotlin/org/aing/danurirest/domain/auth/common/dto/TokenRefreshRequest.kt
@@ -1,4 +1,4 @@
-package org.aing.danurirest.domain.auth.admin.dto
+package org.aing.danurirest.domain.auth.common.dto
 
 import jakarta.validation.constraints.NotNull
